✨ feat(mobile/content): add ContentItemSkeleton for loading states and integrate into ContentDetailModal and ContentListSkeleton
